Improve and fix reporting of CMN configuration registers
A check of all the non-zero registers in the CMN memory space,
against the registers touched by the various reporting scripts
(cmn_list, cmn_dtstat etc.) revealed that some interesting
registers were not read, and some offsets were incorrect.
The reporting scripts are now more complete and correct.
A new script cmn_errstat.py has been added to report on
error-logging messages, which are probably only of specialist
interest. The scripts now report on

Note that there are still gaps in coverage of registers that are
normally Secure-readable only, e.g. custom mesh routing is not yet
covered in reporting.