cmn_discovery.py can now specify CMN location explicitly
Some commands had --cmn-base/--cmn-root-offset options to allow
the CMN location (and, for CMN-600, the root offset) to be specified
explicitly in case it could not be discovered from firmware tables.
cmn_discovery.py lacked these options, which have now been added.
It is still recommended to describe CMN in firmware tables.