dmi: fix a crash caused by empty first string of structure
dmi.py was crashing on some input that had an empty string as
the first string of a structure. As a rule SMBIOS can't handle
empty strings (since an empty string following another string
would appear as the double-nul terminator of a string section)
but dmidecode appeared to handle this by scanning forwards to
the double-nul, so do the same.